+++ src/application/shell.ts
@@ -18,10 +18,6 @@ import {
   StackedPanel, TabBar, Title, Widget
 } from '@phosphor/widgets';
 
-import {
-  IInstanceRestorer
-} from '../instancerestorer';
-
 
 /**
  * The class name added to AppShell instances.
@@ -165,7 +161,7 @@ class ApplicationShell extends Widget {
   /**
    * Promise that resolves when state is restored, returning layout description.
    */
-  get restored(): Promise<ApplicationShell.Hydrated.ILayout> {
+  get restored(): Promise<ApplicationShell.Dehydrated.ILayout> {
     return this._restored.promise;
   }
 
@@ -440,7 +436,7 @@ class ApplicationShell extends Widget {
     }
     const dock = this._dockPanel;
     console.log(dock.saveLayout());
-    let data: ApplicationShell.Hydrated.ILayout = {
+    let data: ApplicationShell.Dehydrated.ILayout = {
       currentWidget: this._tracker.currentWidget,
       leftArea: this._leftHandler.dehydrate(),
       rightArea: this._rightHandler.dehydrate()
@@ -484,7 +480,7 @@ class ApplicationShell extends Widget {
   private _hsplitPanel: SplitPanel;
   private _isRestored = false;
   private _leftHandler: Private.SideBarHandler;
-  private _restored = new PromiseDelegate<ApplicationShell.Hydrated.ILayout>();
+  private _restored = new PromiseDelegate<ApplicationShell.Dehydrated.ILayout>();
   private _rightHandler: Private.SideBarHandler;
   private _topPanel: Panel;
   private _tracker = new FocusTracker<Widget>();
@@ -526,7 +522,7 @@ namespace ApplicationShell {
    * A namespace for hydrated restorable state.
    */
   export
-  namespace Hydrated {
+  namespace Dehydrated {
     /**
      * A description of the application's user interface layout.
      */
@@ -582,60 +578,6 @@ namespace ApplicationShell {
   }
 
   /**
-   * A namespace for dehydrated restore state that can be serialized and saved.
-   */
-  export
-  namespace Dehydrated {
-    /**
-     * The dehydrated state of the application layout.
-     *
-     * #### Notes
-     * This format is JSON serializable and saved in the state database.
-     * It is meant to be a data structure can translate into an
-     * `ApplicationShell.Hydrated.ILayout` data structure for consumption by the
-     * application shell.
-     */
-    export
-    interface ILayout extends JSONObject {
-      /**
-       * The current widget that has application focus.
-       */
-      current?: string | null;
-
-      /**
-       * The left area of the user interface.
-       */
-      left?: ISideArea | null;
-
-      /**
-       * The right area of the user interface.
-       */
-      right?: ISideArea | null;
-    }
-
-    /**
-     * The restorable description of a sidebar in the user interface.
-     */
-    export
-    interface ISideArea extends JSONObject {
-      /**
-       * A flag denoting whether the sidebar has been collapsed.
-       */
-      collapsed?: boolean | null;
-
-      /**
-       * The current widget that has side area focus.
-       */
-      current?: string | null;
-
-      /**
-       * The collection of widgets held by the sidebar.
-       */
-      widgets?: Array<string> | null;
-    }
-  }
-
-  /**
    * An application layout data store.
    */
   export
@@ -647,12 +589,12 @@ namespace ApplicationShell {
      * Fetching the layout relies on all widget restoration to be complete, so
      * calls to `fetch` are guaranteed to return after restoration is complete.
      */
-    fetch(): Promise<ApplicationShell.Hydrated.ILayout>;
+    fetch(): Promise<ApplicationShell.Dehydrated.ILayout>;
 
     /**
      * Save the layout state for the application.
      */
-    save(data: ApplicationShell.Hydrated.ILayout): Promise<void>;
+    save(data: ApplicationShell.Dehydrated.ILayout): Promise<void>;
   }
 }
 
@@ -764,15 +706,8 @@ namespace Private {
 
     /**
      * Dehydrate the side bar data.
-     *
-     * #### Notes
-     * While this method is called `dehydrate`, what it actually returns is a
-     * "live" (hydrated) version of the data structure that defines a side area.
-     * The job of serializing this live object is delegated to the layout
-     * restorer that consumes this object. So as far as the application shell is
-     * concerned, this data structure *is* dehydrated.
      */
-    dehydrate(): ApplicationShell.Hydrated.ISideArea {
+    dehydrate(): ApplicationShell.Dehydrated.ISideArea {
       let collapsed = this._sideBar.currentTitle === null;
       let widgets = toArray(this._stackedPanel.widgets);
       let currentWidget = widgets[this._sideBar.currentIndex];
@@ -782,7 +717,7 @@ namespace Private {
     /**
      * Rehydrate the side bar.
      */
-    rehydrate(data: ApplicationShell.Hydrated.ISideArea): void {
+    rehydrate(data: ApplicationShell.Dehydrated.ISideArea): void {
       if (data.currentWidget) {
         this.activate(data.currentWidget.id);
       } else if (data.collapsed) {
